net/devif_callback: change flags type from uint16_t to uint32_t
increase the size of the flag to prepare for precise xxx_POLL.

+++ net/can/can_recvmsg.c
@@ -288,8 +288,8 @@ static inline int can_readahead(struct can_recvfrom_s *pstate)
   return 0;
 }
 
-static uint16_t can_recvfrom_eventhandler(FAR struct net_driver_s *dev,
-                                          FAR void *pvpriv, uint16_t flags)
+static uint32_t can_recvfrom_eventhandler(FAR struct net_driver_s *dev,
+                                          FAR void *pvpriv, uint32_t flags)
 {
   struct can_recvfrom_s *pstate = pvpriv;
